pyproject.toml
[tool.poetry]
name = "tinkerqa-discord"
version = "0.3.0"
description = ""
authors = ["Robert <xiurobert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
py-cord = {version = "^2.0.0*", allow-prereleases = true}
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
